Add files via upload
Adding first project: 
Project Highlights some of my foundational python skills. (EDA) (Univariate Analysis) (Bi-variate Analysis)

The food aggregator company has stored the data of the different orders made by the registered customers in their online portal. They want to analyze the data to draw some actionable insights for the business. Suppose you are hired as a Data Scientist in this company and the Data Science team has shared some of the key questions that need to be answered. Perform the data analysis to find answers to these questions that will help the company to improve the business.